+++ Leomard/Models/Lemmy/Forms/ModTools/LockPost.swift
@@ -1,0 +1,13 @@
+//
+//  LockPost.swift
+//  Leomard
+//
+//  Created by Konrad Figura on 27/07/2023.
+//
+
+import Foundation
+
+struct LockPost: Codable {
+    let locked: Bool
+    let postId: Int
+}