+++ Leomard/Models/Leomard/LemmyInstance.swift
@@ -10,4 +10,5 @@ import Foundation
 struct LemmyInstance: Hashable {
     let name: String
     let url: String
+    var image: URL?
 }