+++ Leomard/Services/CommentService.swift
@@ -141,4 +141,18 @@ class CommentService: Service {
             self.respond(result, completion)
         }
     }
+    
+    public func distinguish(comment: Comment, distinguished: Bool, completion: @escaping (Result<CommentResponse, Error>) -> Void) {
+        let body = DistinguishComment(commentId: comment.id, distinguished: distinguished)
+        requestHandler.makeApiRequest(host: SessionStorage.getInstance.getLemmyInstance(), request: "/comment/distinguish", method: .post, body: body) { result in
+            self.respond(result, completion)
+        }
+    }
+    
+    public func remove(comment: Comment, removed: Bool, reason: String, completion: @escaping (Result<CommentResponse, Error>) -> Void) {
+        let body = RemoveComment(commentId: comment.id, reason: reason.count > 0 ? reason : nil, removed: removed)
+        requestHandler.makeApiRequest(host: SessionStorage.getInstance.getLemmyInstance(), request: "/comment/remove", method: .post, body: body) { result in
+            self.respond(result, completion)
+        }
+    }
 }