+++ Leomard/Services/SearchService.swift
@@ -113,4 +113,42 @@ class SearchService: Service {
             }
         }
     }
+    
+    public func search(person: Person, query: String, searchType: SearchType, page: Int, completion: @escaping (Result<SearchResponse, Error>) -> Void) {
+        let searchQuery = query.replacingOccurrences(of: " ", with: "%20")
+        let request = "/search?q=\(searchQuery)&type_=\(String(describing: searchType))&page=\(String(page))&creator_id=\(person.id)"
+        requestHandler.makeApiRequest(host: SessionStorage.getInstance.getLemmyInstance(), request: request, method: .get) { result in
+            switch result {
+            case .success(let apiResponse):
+                do {
+                    if let data = apiResponse.data {
+                        var response = try self.decode(type: SearchResponse.self, data: data)
+                        response.posts.forEach { postView in
+                            if postView.post.nsfw && !UserPreferences.getInstance.showNsfw  {
+                                response.posts = response.posts.filter { $0 != postView }
+                            }
+                        }
+                        
+                        response.communities.forEach { communityView in
+                            if communityView.community.nsfw && !UserPreferences.getInstance.showNsfw {
+                                response.communities = response.communities.filter { $0 != communityView }
+                            }
+                        }
+                        
+                        response.comments.forEach { commentView in
+                            if commentView.post.nsfw && !UserPreferences.getInstance.showNsfw  {
+                                response.comments = response.comments.filter { $0 != commentView }
+                            }
+                        }
+                        
+                        completion(.success(response))
+                    }
+                } catch {
+                    self.respondError(apiResponse.data!, completion)
+                }
+            case .failure(let error):
+                completion(.failure(error))
+            }
+        }
+    }
 }