+++ Leomard/Views/CommunityUIView.swift
@@ -7,9 +7,10 @@
 
 import Foundation
 import SwiftUI
+import MarkdownUI
 
 struct CommunityUIView: View {
-    let community: Community
+    @State var community: Community
     let postService: PostService
     let commentService: CommentService
     let contentView: ContentView
@@ -40,6 +41,22 @@ struct CommunityUIView: View {
     
     @State var isLoading: Bool = false
     
+    @State var showCommunityEdit: Bool = false
+    @State var title: String = ""
+    @State var description: String = ""
+    @State var nsfw: Bool = false
+    @State var postingRestrictedToMods: Bool = false
+    @State var communityUpdateFail: Bool = false
+    
+    @State var imageUploadFail: Bool = false
+    @State var imageUploadFailReason: String = ""
+    
+    @State var showCommunityRemove: Bool = false
+    @State var communityRemoved: Bool = false
+    @State var communityRemoveText: String = ""
+    
+    @Environment(\.openURL) var openURL
+    
     var body: some View {
         toolbar
             .frame(
@@ -62,6 +79,8 @@ struct CommunityUIView: View {
                 )
             }
         }
+        .overlay(communityEditor)
+        .overlay(removeCommunityOverlay)
         .cornerRadius(8)
         .task {
             let requestHandler = RequestHandler()
@@ -69,49 +88,126 @@ struct CommunityUIView: View {
             self.selectedBrowseOption = browseOptions[0]
             loadCommunity()
         }
+        
+        .alert("Community Removed", isPresented: $communityRemoved, actions: {
+            Button("OK", action: {})
+        }, message: {
+            Text("Community has been removed successfully.")
+        })
         Spacer()
     }
     
     // MARK: -
     
     @ViewBuilder
-    private func communityContent(_ communityResponse: GetCommunityResponse?, sidebarVisible: Bool) -> some View {
-        ScrollViewReader { scrollProxy in
-            List {
-                if let communityResponse {
-                    if sidebarVisible {
-                        VStack {
-                            CommunityUISidebarView(
-                                communityResponse: communityResponse,
-                                communityService: communityService!,
-                                contentView: contentView,
-                                myself: $myself,
-                                onPostAdded: addNewPost)
+    private var removeCommunityOverlay: some View {
+        if showCommunityRemove {
+            ZStack {
+                Color(white: 0, opacity: 0.33)
+                    .onTapGesture {
+                        showCommunityEdit = false
+                    }
+                VStack(alignment: .center) {
+                    VStack(alignment: .center) {
+                        Image(nsImage: NSApplication.shared.applicationIconImage)
+                            .resizable()
+                            .aspectRatio(contentMode: .fit)
+                            .frame(width: 100, height: 100)
+                        Text("Remove Community")
+                            .bold()
+                            .font(.title2)
+                    }
+                    .padding()
+                    VStack(alignment: .center) {
+                        Text("Are you really, REALLY sure you want to remove this community? To remove the community, type the following:\n")
+                        Text("**Yes, remove \(community.name)@\(LinkHelper.stripToHost(link: community.actorId)) community.**")
+                            .frame(alignment: .center)
+                            .bold()
+                    }
+                    .frame(maxWidth: .infinity)
+                    TextField("", text: $communityRemoveText)
+                    HStack(alignment: .center) {
+                        Button("Remove Community") {
+                            communityService?.remove(community: community, removed: true) { result in
+                                switch result {
+                                case .success(_):
+                                    communityRemoved = true
+                                case .failure(let error):
+                                    print(error)
+                                }
+                            }
+                            showCommunityRemove = false
                         }
-                        .frame(
-                            minWidth: 0,
-                            maxWidth: .infinity,
-                            alignment: .center
-                        )
-                        .cornerRadius(8)
-                        .padding(.bottom, 15)
+                        .buttonStyle(.plain)
+                        .foregroundColor(.red)
+                        .disabled(communityRemoveText != "Yes, remove \(community.name)@\(LinkHelper.stripToHost(link: community.actorId)) community.")
+                        Button("Cancel") {
+                            showCommunityRemove = false
+                        }
+                        .buttonStyle(.borderedProminent)
                     }
-                    switch selectedBrowseOption.id {
-                    case 1:
-                        commentsList
-                    default:
-                        postsList
-                    }
-                    
                 }
+                .frame(maxWidth: 300)
+                .padding()
+                .background(Color(.windowBackgroundColor))
+                .overlay(
+                    RoundedRectangle(cornerRadius: 16)
+                        .stroke(Color(.windowFrameTextColor), lineWidth: 1)
+                )
+                .listStyle(SidebarListStyle())
+                .scrollContentBackground(.hidden)
+                .task {
+                    communityRemoveText = ""
+                }
+                .cornerRadius(16)
             }
-            .frame(
-                minWidth: 0,
-                maxWidth: 600,
-                maxHeight: .infinity,
-                alignment: .center
-            )
         }
+    }
+    
+    // MARK: -
+    
+    @ViewBuilder
+    private func communityContent(_ communityResponse: GetCommunityResponse?, sidebarVisible: Bool) -> some View {
+        List {
+            if let communityResponse {
+                if sidebarVisible {
+                    VStack {
+                        CommunityUISidebarView(
+                            communityResponse: communityResponse,
+                            communityService: communityService!,
+                            contentView: contentView,
+                            myself: $myself,
+                            onPostAdded: addNewPost,
+                            onEditCommunity: {
+                                showCommunityEdit = true
+                            },
+                            onRemoveCommunity: {
+                                showCommunityRemove = true
+                            })
+                    }
+                    .frame(
+                        minWidth: 0,
+                        maxWidth: .infinity,
+                        alignment: .center
+                    )
+                    .cornerRadius(8)
+                    .padding(.bottom, 15)
+                }
+                switch selectedBrowseOption.id {
+                case 1:
+                    commentsList
+                default:
+                    postsList
+                }
+                
+            }
+        }
+        .frame(
+            minWidth: 0,
+            maxWidth: 600,
+            maxHeight: .infinity,
+            alignment: .center
+        )
     }
     
     @ViewBuilder
@@ -120,7 +216,12 @@ struct CommunityUIView: View {
             List {
                 VStack {
                     if communityResponse != nil {
-                        CommunityUISidebarView(communityResponse: communityResponse!, communityService: communityService!, contentView: contentView, myself: $myself, onPostAdded: addNewPost)
+                        CommunityUISidebarView(communityResponse: communityResponse!, communityService: communityService!, contentView: contentView, myself: $myself, onPostAdded: addNewPost, onEditCommunity: {
+                            showCommunityEdit = true
+                        },
+                                               onRemoveCommunity: {
+                            showCommunityRemove = true
+                        })
                     }
                 }
                 .frame(
@@ -274,6 +375,103 @@ struct CommunityUIView: View {
         }
     }
     
+    @ViewBuilder
+    private var communityEditor: some View {
+        if showCommunityEdit {
+            ZStack {
+                Color(white: 0, opacity: 0.33)
+                    .onTapGesture {
+                        showCommunityEdit = false
+                    }
+                VStack {
+                    VStack(alignment: .leading) {
+                        HStack {
+                            Button("Dismiss", action: { showCommunityEdit = false })
+                                .buttonStyle(.link)
+                            Spacer()
+                            Button("Change More Settings", action: {
+                                self.openURL(URL(string: community.actorId)!)
+                            })
+                            .buttonStyle(.link)
+                        }
+                        .frame(
+                            minWidth: 0,
+                            maxWidth: .infinity,
+                            alignment: .leading
+                        )
+                        .padding(.top, 10)
+                        .padding(.bottom, 0)
+                        Spacer()
+                        
+                        List {
+                            VStack(alignment: .leading) {
+                                Text("Display Name")
+                                    .bold()
+                                TextField("Optional", text: $title)
+                            }
+                            VStack(alignment: .leading) {
+                                Text("Description")
+                                    .bold()
+                                Button("Add Image", action: addImage)
+                                TextEditor(text: $description)
+                                    .overlay(RoundedRectangle(cornerRadius: 8).stroke(.primary, lineWidth: 0.5))
+                                    .frame(
+                                        maxWidth: .infinity,
+                                        minHeight: 3 * NSFont.preferredFont(forTextStyle: .body).xHeight,
+                                        maxHeight: .infinity,
+                                        alignment: .leading
+                                    )
+                                    .lineLimit(5...)
+                                    .font(.system(size: NSFont.preferredFont(forTextStyle: .body).pointSize))
+                                Text("Description Preview")
+                                    .bold()
+                                Markdown(MarkdownContent(description))
+                            }
+                            VStack(alignment: .leading) {
+                                Toggle("NSFW", isOn: $nsfw)
+                                Toggle("Posting Restricted To Moderators", isOn: $postingRestrictedToMods)
+                            }
+                            VStack(alignment: .leading) {
+                                Button("Save Changes", action: {
+                                    saveCommunitySettings()
+                                })
+                                .buttonStyle(.borderedProminent)
+                            }
+                        }
+                        .frame(maxHeight: .infinity)
+                        .textFieldStyle(.roundedBorder)
+                        .task {
+                            self.title = community.title
+                            self.description = community.description ?? ""
+                            self.nsfw = community.nsfw
+                            self.postingRestrictedToMods = community.postingRestrictedToMods
+                        }
+                    }
+                    .frame(maxWidth: .infinity, maxHeight: .infinity)
+                }
+                .padding()
+                .frame(maxWidth: 600, maxHeight: 600)
+                .background(Color(.textBackgroundColor))
+                .overlay(
+                    RoundedRectangle(cornerRadius: 4)
+                        .stroke(Color(.windowFrameTextColor), lineWidth: 1)
+                )
+                .listStyle(SidebarListStyle())
+                .scrollContentBackground(.hidden)
+            }
+            .alert("Image Upload Failed", isPresented: $imageUploadFail, actions: {
+                Button("OK") {}
+            }, message: {
+                Text(imageUploadFailReason)
+            })
+            .alert("Community Update Failure", isPresented: $communityUpdateFail, actions: {
+                Button("OK") {}
+            }, message: {
+                Text("Failed to update the community. Try again later.")
+            })
+        }
+    }
+    
     // MARK: -
     
     func loadCommunity() {
@@ -303,7 +501,7 @@ struct CommunityUIView: View {
             isLoading = false
             switch result {
             case .success(let postsResponse):
-                self.posts += postsResponse.posts
+                self.posts += postsResponse.posts.filter { !self.posts.contains($0) }
                 page += 1
             case .failure(let error):
                 print(error)
@@ -317,7 +515,7 @@ struct CommunityUIView: View {
             isLoading = false
             switch result {
             case .success(let commentsResponse):
-                self.comments += commentsResponse.comments
+                self.comments += commentsResponse.comments.filter { !self.comments.contains($0) }
                 page += 1
             case .failure(let error):
                 print(error)
@@ -388,10 +586,64 @@ struct CommunityUIView: View {
             isLoading = false
             switch result {
             case .success(let response):
-                self.posts += response.posts
+                self.posts += response.posts.filter { !self.posts.contains($0) }
                 page += 1
             case .failure(let error):
                 print(error)
+            }
+        }
+    }
+    
+    func addImage() {
+        let panel = NSOpenPanel()
+        panel.prompt = "Select file"
+        panel.allowsMultipleSelection = false
+        panel.canChooseDirectories = false
+        panel.canCreateDirectories = false
+        panel.canChooseFiles = true
+        panel.allowedContentTypes = [
+            .init(importedAs: "leomard.supported.image.types.jpg"),
+            .init(importedAs: "leomard.supported.image.types.jpeg"),
+            .init(importedAs: "leomard.supported.image.types.png"),
+            .init(importedAs: "leomard.supported.image.types.webp"),
+            .init(importedAs: "leomard.supported.image.types.gif")
+        ]
+        panel.begin { (result) -> Void in
+            self.contentView.toggleInteraction(true)
+            if result.rawValue == NSApplication.ModalResponse.OK.rawValue, let url = panel.url {
+                let imageService = ImageService(requestHandler: RequestHandler())
+                imageService.uploadImage(url: url) { result in
+                    switch result {
+                    case .success(let imageUploadResponse):
+                        if description.count > 0 {
+                            description += "\n\n"
+                        }
+                        
+                        description += "![](\(imageUploadResponse.data.link))\n\n"
+                    case .failure(let error):
+                        if error is LeomardExceptions {
+                            self.imageUploadFailReason = String(describing: error as! LeomardExceptions)
+                        } else {
+                            self.imageUploadFailReason = "Unable to upload the image :("
+                        }
+                        
+                        self.imageUploadFail = true
+                    }
+                }
+            }
+        }
+        panel.orderFrontRegardless()
+        self.contentView.toggleInteraction(false)
+    }
+    
+    func saveCommunitySettings() {
+        communityService?.edit(community: community, displayName: title, description: description, nsfw: nsfw, postingRestrictedToMods: postingRestrictedToMods) { result in
+            switch result {
+            case .success(let communityResponse):
+                self.community = communityResponse.communityView.community
+            case .failure(let error):
+                print(error)
+                communityUpdateFail = true
             }
         }
     }