+++ Leomard/Views/NavbarView.swift
@@ -21,6 +21,11 @@ struct NavbarView: View {
     @State var emptyCounter: Int = 0
     @State var followedVisible: Bool = true
     
+    @State var addedLetters: [Character] = []
+    
+    @State var searchQuery: String = ""
+    @State var searchVisible: Bool = false
+    
     var body: some View {
         sidebarItems(options: options)
             .padding(.top, 24)
@@ -31,7 +36,7 @@ struct NavbarView: View {
             )
         userProfileItem
     }
-
+    
     // MARK: -
     
     @ViewBuilder
@@ -46,22 +51,61 @@ struct NavbarView: View {
     
     @ViewBuilder
     private var followedCommunitiesList: some View {
-        List {
-            HStack(spacing: 14) {
-                Text("Followed")
-                Image(systemName: "v.circle")
-                    .foregroundColor(.secondary)
-                    .rotationEffect(.degrees(followedVisible ? 180 : 0))
-            }
-            .onTapGesture {
-                followedVisible = !followedVisible
-            }
-            if followedVisible {
-                ForEach(followedCommunities, id: \.self) { communityView in
-                    NavbarCommunityItem(community: communityView.community, currentCommunity: $currentCommunity, contentView: contentView)
+        VStack {
+            VStack(alignment: .leading) {
+                HStack {
+                    HStack(spacing: 14) {
+                        Text("Followed")
+                        Image(systemName: "v.circle")
+                            .foregroundColor(.secondary)
+                            .rotationEffect(.degrees(followedVisible ? 180 : 0))
+                    }
+                    .onTapGesture {
+                        followedVisible = !followedVisible
+                    }
+                    Image(systemName: "magnifyingglass")
+                        .foregroundColor(searchVisible ? Color(.linkColor) : .secondary)
+                        .onTapGesture {
+                            searchVisible = !searchVisible
+                            if !searchVisible {
+                                searchQuery = ""
+                            }
+                        }
+                }
+                .frame(maxWidth: .infinity, alignment: .leading)
+                if searchVisible {
+                    TextField("Search", text: $searchQuery)
+                        .textFieldStyle(.roundedBorder)
                 }
             }
+            .padding(.leading)
+            .padding(.trailing)
+            .frame(maxWidth: .infinity, alignment: .leading)
+            List {
+                if followedVisible {
+                    ForEach(followedCommunities.sorted(by: {
+                        UserPreferences.getInstance.preferDisplayNameCommunityFollowed ? $0.community.title < $1.community.title
+                        : $0.community.name < $1.community.name
+                    }).filter {
+                        if searchQuery.count > 0 {
+                            return $0.community.name.contains(searchQuery) || $0.community.title.contains(searchQuery)
+                        } else {
+                            return true
+                        }
+                    }
+                            , id: \.self) { communityView in
+                        if UserPreferences.getInstance.navbarShowLetterSeparators,
+                           let firstChar =  UserPreferences.getInstance.preferDisplayNameCommunityFollowed ? communityView.community.title.first : communityView.community.name.first,
+                           isFirstCommunityStartingWithThisChar(community: communityView.community) {
+                            Text(String(firstChar.uppercased()))
+                        }
+                        NavbarCommunityItem(community: communityView.community, currentCommunity: $currentCommunity, contentView: contentView)
+                    }
+                }
+            }
+            .frame(maxHeight: .infinity)
         }
+        .frame(maxHeight: .infinity)
     }
     
     @ViewBuilder
@@ -71,4 +115,8 @@ struct NavbarView: View {
         }
         .NavBarItemContainer()
     }
+    
+    func isFirstCommunityStartingWithThisChar(community: Community) -> Bool {
+        return followedCommunities.filter { $0.community.name.first!.uppercased() == community.name.first!.uppercased() }.first?.community == community
+    }    
 }